BCSSSegmentationContainer/requirements.txt
hooknet
SimpleITK
scikit-image
numpy
Pillow
matplotlib
evalutils
